+++ storage/configsetup.py
@@ -10,10 +10,18 @@ config = configparser.ConfigParser()
 config.read("config.ini")
 
 # Check our working directory. This script can't work in the base PyWeather folder.
-
+        
 if "/pyweather/storage" in os.getcwd():
     cool = True
 else:
+    # If we are in the base folder, it's a perfect time to write to versioninfo.txt!
+    try:
+        versioninfo = open('updater//versioninfo.txt').close()
+    except:
+        open('updater//versioninfo.txt', 'w').close()
+        with open("updater//versioninfo.txt", 'a') as out:
+            out.write("0.6 beta")
+            out.close()
     print("Whoops! You're running this script from PyWeather's base folder, but",
           "the script can't work when running from the base folder of PyWeather.",
           "Move into the storage folder, and run the script there.",